Fixed chunks calculation
The chunks were calculated by dividing by MESH_PAYLOAD_SIZE but this
gave one extra chunk when it was an exact multiple. The correct equation
is to first subract 1 from the total size.

Also added option to manually disconnect a node from the broker and view
its listed peers.

The test for the last chunk was moved to when the chunk is identified as
being a contiuation of a previous one.

The Mesh SSID Password and Key commands have been changed to use " to
clear them. An empty message now returns the set values.