Improved packet forwarding
Packets received with a TTL of 1 are only forwarded to the receiver if
it is known to the node.

Time packets are always forwarded even if the node was the receiver.
When a node requests a time packet it is set as the receiver of time
packets.

Also expanded MeshNode to give the role when no argument is provided.